+++ src/main/java/cc/cassian/clickthrough/config/ModConfig.java
@@ -1,61 +1,20 @@
 package cc.cassian.clickthrough.config;
 
-import cc.cassian.clickthrough.ClickThrough;
-import com.google.gson.FieldNamingPolicy;
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import dev.architectury.injectables.annotations.ExpectPlatform;
+import folk.sisby.kaleido.api.WrappedConfig;
+import folk.sisby.kaleido.lib.quiltconfig.api.values.ValueList;
 
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.OutputStreamWriter;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
+import java.util.ArrayList;
 import java.util.List;
 
-public class ModConfig {
+public class ModConfig extends WrappedConfig {
 
-    private static final Gson GSON = new GsonBuilder().setPrettyPrinting().setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES).create();
-
-    private static ModConfig INSTANCE = new ModConfig();
     //General settings
     public int version = 0;
     public boolean isActive = true;
     public boolean onlycontainers = true;
     public boolean sneaktodye = false;
-    public List<String> containers = List.of("minecraft:ender_chest", "minecraft:vault", "minecraft:composter", "minecraft:respawn_anchor", "minecraft:jukebox", "minecraft:decorated_pot", "minecraft:chiseled_bookshelf", "minecraft:beacon", "minecraft:stonecutter", "minecraft:grindstone", "minecraft:crafting_table");
+    public ValueList<String> containers = ValueList.create("minecraft:ender_chest", "minecraft:vault", "minecraft:composter", "minecraft:respawn_anchor", "minecraft:jukebox", "minecraft:decorated_pot", "minecraft:chiseled_bookshelf", "minecraft:beacon", "minecraft:stonecutter", "minecraft:grindstone", "minecraft:crafting_table");
 
-    public static void load() {
-        if (!Files.exists(configPath())) {
-            save();
-            return;
-        }
 
-        try (var input = Files.newInputStream(configPath())) {
-            INSTANCE = GSON.fromJson(new InputStreamReader(input, StandardCharsets.UTF_8), ModConfig.class);
-        } catch (IOException e) {
-            ClickThrough.LOGGER.warn("Unable to load config file!");
-        }
-    }
-
-    public static void save() {
-        try (var output = Files.newOutputStream(configPath()); var writer = new OutputStreamWriter(output, StandardCharsets.UTF_8)) {
-            GSON.toJson(INSTANCE, writer);
-        } catch (IOException e) {
-            ClickThrough.LOGGER.warn("Unable to save config file!");
-        }
-        ModLists.loadLists();
-    }
-
-    public static ModConfig get() {
-        if (INSTANCE == null) INSTANCE = new ModConfig();
-        return INSTANCE;
-    }
-
-    @ExpectPlatform
-    static Path configPath() {
-        throw new AssertionError();
-    }
 
 }